requirements.txt
openai>=1.51.0
python-dotenv>=1.0.1
colorama>=0.4.6
streamlit>=1.37.0
flask>=3.0.3
authlib>=1.3.2
SQLAlchemy>=2.0.36
